fix: 객체타입 수정

+++ back/src/main/java/com/back/domain/order/order/dto/OrderResponseDto.java
@@ -1,12 +1,8 @@
 package com.back.domain.order.order.dto;
 
 
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-
 import com.back.domain.order.order.entity.Order;
 import com.back.domain.orderItem.orderItem.entity.OrderItem;
-
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -35,7 +31,7 @@ public record OrderResponseDto(Long id,
 
         // 혹시 long으로 강제cast를 하신 이유가 있을까요??
         return new OrderResponseDto(
-                (long) order.getId(),
+                order.getId(),
                 order.getEmail(),
                 order.getAddress(),
                 items,
